// src/exam_lists/dto/create-exam_lists.dto.ts
export class CreateExam_listDto {
  group_id: number;
  position: number;
  description: string;
  size: number;
  annulled: boolean;
  abbreviation: string;
  special_test: boolean;
  work_sheet: string;
  tax_id: number;
  format: string;
  format_vue: string;
  cost1: number;
  cost2: number;
  cost3: number;
  cost4: number;
  cost5: number;
  cost6: number;
}